this is more of a temporary optimization than a final solution.
the basic problem is that the pool size for reference counting proxy
objects grows linearly. if you suddenly need millions of reference counted
pointers, you spend minutes in realloc. this is the sad times.
making the default pool size way larger effectively makes the slope
of the line steep enough to not cause problems for my use case.
the better solution would be to do something like grow the pool
non-linearly, but I'm not sure I understand this codebase well enough
to make that not break everything.